Rework the SeachBy API
Currently the way you specify what to search by is to call the correct
function. This changes things so that there is only one SearchBy
function and it accepts a second argument to specify what to search by.

This makes the code clean but also more flexable for clients to use.

For example a tool where SearchBy is intended to be user configuable.
The SearchBy can be stored in some config struct and then just passed to
the calls to SearchBy.